requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
Pillow==8.4.0
PyPDF2==1.26.0